pyproject.toml
[project]
name = "twilio-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.9.1",
    "fastapi==0.104.1",
    "psutil>=7.0.0",
    "pydantic==2.4.2",
    "python-dotenv==1.0.0",
    "uvicorn[standard]==0.24.0",
    "websockets==12.0",
]
